Updating to version 1.1b
Two extra buttons allow to switch between "all cookies" and "flagged cookies".
"All cookies" show all cookies which are active for the site at the moment.
"Flagged cookies" list also cookies which are set on the domain, if those have been set to flagged. If they are not active, they are set to the end of the list and only shown with name, but not with there values (which are not stored!)