pdate to version 1.2
This version has been restyled and should be more pleasent to use, this was done by help from @Cybergreek at https://ngb.to !

"All cookies" has been renamed to "Active cookies" as those are the cookies, which are currently active for the domain.
"Flagged cookies" still lists "all flagged cookies for the domain", but finally has the option to remove cookies, which are not active at preset, but set to be flagged, which otherwise would be hidden.

Also a search is included, which filters "Active cookies" and "Flagged cookies" easily.