Bugfix release 1.4a
- Unflagging cookies from "flagged cookies" resulted in left over "flagged state" for the cookies, this is resolved.
The "auto flag" feature also confused the state and action and set it to the next level and permitted those cookies then.
So by now, "unchecking" any flagged cookie from the "flagged cookie list" will correctly remove the flagging for the cookie and delete its stored value.
The functions have been seperated, before it was shared, leading also to some overhead which was not required.

- Added some more optimizations to the code and also using 'standard' as the main syntax