Changing HTML information
Earlier, the display was labeled "active" - for active domain cookies, which was correct.
At the moment of writing, all cookies, which are found for a domain are listed there, not anymore alone active cookies.
This is due a problem that cookies are erased, before they really had appeared in the control UI.